+ added load to cluster table ~ fixed nodeIDs not being the right type ~ added back some session variable get/set to update side-effects too ~ fixed inventory accessing as everything was being loaded instead of the given flags ~ reworked SystemManager to properly handle loading and discovery of solar systems + added support for sesion change notifications on nodes, so updated data can be fetched + added support for broadcast packets (kind of rudimentary but it works) + added hearbeat support to nodes/proxies + added permission checking for service calls to prevent calls when a service requires certain client's state

+++ Database/Server/04-entity.sql
@@ -19,7 +19,7 @@ CREATE TABLE `invItems` (
   `singleton` int(10) unsigned NOT NULL default '0',
   `quantity` int(10) unsigned NOT NULL default '0',
   `customInfo` text,
-  `nodeID` int(10) unsigned DEFAULT NULL,
+  `nodeID` bigint(20) DEFAULT NULL,
   PRIMARY KEY  (`itemID`),
   KEY `typeID` (`typeID`),
   KEY `locationID` (`locationID`),
